Implement panel-scale RTL DXA stream

+++ tests/regression/coop_dxa_lmem_write/common.h
@@ -10,17 +10,17 @@ typedef struct {
   uint64_t src0_addr;
   uint64_t src1_addr;
   uint64_t dst_addr;
-  uint32_t expected_slot0;
-  uint32_t expected_slot1;
+  uint64_t expected_slot0;
+  uint64_t expected_slot1;
 } kernel_arg_t;
 
 typedef struct {
   uint32_t team_rank_x;
   uint32_t team_rank_y;
-  uint32_t slot0_value;
-  uint32_t slot1_value;
-  uint32_t expected_slot0;
-  uint32_t expected_slot1;
+  uint64_t slot0_value;
+  uint64_t slot1_value;
+  uint64_t expected_slot0;
+  uint64_t expected_slot1;
   uint32_t status;
 } result_t;
